requirements.txt
django==3.2.7
djangorestframework==3.12.2
python-dotenv==0.19.0
whitenoise==5.3.0
pillow==8.3.2
django-tinymce==3.3.0
drf_yasg==1.20.0